+++ packages/k8s-novolume/tests/run-container-step-test.ts
@@ -42,7 +42,7 @@ describe('Run container step', () => {
           {
             name: JOB_CONTAINER_EXTENSION_NAME,
             command: ['sh'],
-            args: ['-c', 'echo test']
+            args: ['-c', 'sleep 10000']
           },
           {
             name: 'side-container',
@@ -61,10 +61,6 @@ describe('Run container step', () => {
     await expect(
       runContainerStep(runContainerStepData.args)
     ).resolves.not.toThrow()
-    delete process.env[ENV_HOOK_TEMPLATE_PATH]
-  })
-
-  afterEach(() => {
     delete process.env[ENV_HOOK_TEMPLATE_PATH]
   })