+++ releaseNotes.md
@@ -23,3 +23,4 @@ The SHA-256 checksums for the packages included in this build are shown below:
 
 - actions-runner-hooks-docker-<HOOK_VERSION>.zip <DOCKER_SHA>
 - actions-runner-hooks-k8s-<HOOK_VERSION>.zip <K8S_SHA>
+- actions-runner-hooks-k8s-novolume-<HOOK_VERSION>.zip <K8S_NOVOLUME_SHA>